feat: move evm store to stores dir
