+++ apps/trading/stores/evm/index.ts
@@ -1,0 +1,5 @@
+export { type TxDeposit } from './deposit-slice';
+export { type TxSquidDeposit } from './squid-deposit-slice';
+export { type TxWithdraw } from './withdraw-slice';
+export { type TxFaucet } from './faucet-slice';
+export { useEvmTxStore, type Tx } from './evm';
